Chore/rename dscp to sequence (#47)
* chore: Rename dscp to sqnc

First attempt at renaming DSCP to Sequence. Note this also will replace the existing flux infra with one using L3 as that's more indicative of current Sequence deployments.

* chore: Remove ory from kind cluster

We no longer use ory to handle authentication and therefore this does not represent an example part of kind infrastructure. Removing in preparation for keycloak

* Change sync branch

* Attempt at upgrading flux

* Remove secret as flux repo is public. Fix values yaml path

* Modify api values.yaml

* Add pathType

* Update values for matchmaker

* Fix values again

* Fix values key

* Auth type none for kind cluster

* trace logging

* Update tag for debugging

* remove node version overrides

* revert branch to main

* Remove image tag